# backend/ent/schema/room.go
package schema

import (
	"github.com/facebookincubator/ent"
	"github.com/facebookincubator/ent/schema/edge"
)

// Room holds the schema definition for the Room entity.
type Room struct {
	ent.Schema
}

// Fields of the Room.
func (Room) Fields() []ent.Field {
	return nil
}

// Edges of the Room.
func (Room) Edges() []ent.Edge {
	return []ent.Edge{
		edge.From("user", User.Type).Ref("user2room").Unique(),
		edge.To("room2problem", Problem.Type).StorageKey(edge.Column("room_id")),
	}
}